requirements.txt
flask
google-generativeai
python-multipart
python-dotenv
